close #75 Add a 'Next Case' button that will move on to the next case.
close #74 Implemented a 'Discard Case' button
instead. It sets all the items to 'Not Applicable'
and then moves to the 'Next Case'.  Did not want
to make the user have to unset all the values to
be able to discard it.  They should be able to
discard the whole thing at any moment.  Dropped
the heavy validation.